Refactor: Remove Terms of Service and Privacy Policy files; integrate new Privacy Policy and Terms of Service implementations
- Deleted old Terms of Service implementation.
- Deleted old Privacy Policy implementation.
- Added new Privacy Policy with comprehensive user data handling guidelines.
- Added new Terms of Service with updated user agreement and liability clauses.
- Updated Settings viewmodel to reference new Privacy Policy and Terms of Service.
- Improved tag counting functionality in Home viewmodel using isolates for performance.
- Cleaned up unused imports in challenges and near_me viewmodels.
- Updated pubspec.lock to reflect dependency upgrades and changes.
