+++ flutter_app/lib/services/chat_gpt_service.dart
@@ -142,5 +142,5 @@ class ChatGPTService {
 
 String summarizeResponse(String fullText) {
   if (fullText.trim().isEmpty) return "No response to summarize.";
-  return fullText.split('.').first.trim() + '...';
+  return '${fullText.split('.').first.trim()}...';
 }
