+++ flutter_app/lib/views-viewmodels/account/Log_Out.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 class LogoutScreen extends StatelessWidget {
   const LogoutScreen({super.key});
