+++ flutter_app/lib/views-viewmodels/points/subtabs/challenges.dart
@@ -4,7 +4,6 @@ import 'dart:async';
 import 'dart:io';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:ecosphere/repositories/challenges_repository.dart';
-import 'package:ecosphere/services/connectivity.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:shared_preferences/shared_preferences.dart';
